render: calibrate border line weight + cell/table spacing to the original
Gov-doc tables are crisp hairlines, but our borders read heavier/uneven than
the原본 양식 PDF because the weight pipeline had three defects:

1. rhwp `border_width_to_px` rounded every width up to ≥1px, collapsing the two
   thinnest indices (0.4/0.5px) that gov-doc tables use into one heavier 1px line.
   → return f64 and clamp only the floor (0.5px hairline); sub-px detail survives.
   IR `CellEdge.width_px`/`CellDiagonal.width_px` are now f64 (drop the unused Eq;
   PartialEq still drives jsx value-equality; the jsx codec round-trips f64).
2. SVG sink ran the already-px Line width through px() (÷75), collapsing EVERY
   stroke to the 0.4px floor regardless of index. → use the px width directly,
   clamped to a shared BORDER_HAIRLINE_PX (0.5px).
3. PDF per-edge lines floored at 0.3pt (0.75pt/px) while the legacy box stroke was
   0.375pt — a 2× mismatch between cell types. → both now share BORDER_HAIRLINE_PX
   (0.5px → 0.375pt), so legacy-box and per-edge cells print at the same weight.

Verified by exporting 예비창업패키지.hwp and reading pages 1/2/4 vs the original:
borders + the page-4 pentagon banner + dotted guide-box now read as uniform crisp
hairlines matching the original's weight. Cell padding / vertical-centering already
matched the original, so spacing is left untouched to avoid regressing pagination.

Oracle unchanged: 쪽수 8==8, 줄수 정확 일치 98.9%. Added tests: the width-index
mapping (distinct hairlines + floor clamp) and the placement-layer sub-floor clamp.

+++ crates/hwp-model/src/document.rs
@@ -318,12 +318,13 @@ pub struct Cell {
 }
 
 /// One cell edge's rendered border (a side of the box). Lifted from a borderFill `BorderLine`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub struct CellEdge {
     pub color: crate::types::Color,
     pub style: LineStyle,
     /// Stroke width in device px (≈ the SVG/PDF stroke width). Lifted from the HWP width index.
-    pub width_px: u32,
+    /// f64 so sub-px gov-doc hairlines (0.5/0.7px) survive — not rounded up to a heavier 1px.
+    pub width_px: f64,
 }
 
 /// How a border line is drawn. `None` = 선없음 (the edge is suppressed: NO stroke emitted) — this is
@@ -340,11 +341,12 @@ pub enum LineStyle {
 }
 
 /// A cell diagonal line (HWP borderFill `diagonal`). `kind` picks which corners it connects.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub struct CellDiagonal {
     pub kind: DiagonalKind,
     pub color: crate::types::Color,
-    pub width_px: u32,
+    /// Stroke width in device px (f64 — same hairline rationale as [`CellEdge::width_px`]).
+    pub width_px: f64,
 }
 
 /// Slash = bottom-left→top-right (/); BackSlash = top-left→bottom-right (\\).
